feat: :alembic: Send Mail add to schedule to day last hour
PHPMail library has been added to be able to generate emails from the samebit system, with the initial purpose of sending a report at the end of the day with the records created by the APH in order to have a more accurate control of what is done during the day. This will be done at noon and at the end of the day for a more precise time control.